+++ src/idc/imgstats/writer/_area_histogram.py
@@ -9,6 +9,7 @@ from typing import List, Dict
 
 from wai.logging import LOGGING_WARNING
 
+from seppl import PlaceholderSupporter, placeholder_list
 from idc.api import ObjectDetectionData, ImageSegmentationData, StreamWriter, \
     make_list, LABEL_KEY
 
@@ -22,7 +23,7 @@ OUTPUT_FORMATS = [
 ]
 
 
-class AreaHistogramWriter(StreamWriter):
+class AreaHistogramWriter(StreamWriter, PlaceholderSupporter):
     """
     Collects the labels and outputs their distribution.
     """
@@ -89,7 +90,7 @@ class AreaHistogramWriter(StreamWriter):
         :rtype: argparse.ArgumentParser
         """
         parser = super()._create_argparser()
-        parser.add_argument("-o", "--output_file", type=str, help="The file to write the statistics to; uses stdout if omitted.", required=False, default=None)
+        parser.add_argument("-o", "--output_file", type=str, help="The file to write the statistics to; uses stdout if omitted. " + placeholder_list(obj=self), required=False, default=None)
         parser.add_argument("-f", "--output_format", type=str, help="The format to use for the output, available modes: %s" % ", ".join(OUTPUT_FORMATS), required=False, default=OUTPUT_FORMAT_TEXT)
         parser.add_argument("-k", "--label_key", type=str, help="The key in the (object detection) meta-data that contains the label.", required=False, default=LABEL_KEY)
         parser.add_argument("-B", "--num_bins", type=int, help="The number of bins to use for the histogram.", required=False, default=20)
@@ -199,7 +200,7 @@ class AreaHistogramWriter(StreamWriter):
         if use_stdout:
             print(plots)
         else:
-            with open(self.output_file, "w") as fp:
+            with open(self.session.expand_placeholders(self.output_file), "w") as fp:
                 fp.write(plots)
 
     def output_csv(self, histograms: Dict, keys: List[str], use_stdout: bool):
@@ -217,7 +218,7 @@ class AreaHistogramWriter(StreamWriter):
             writer = csv.writer(sys.stdout)
             f = None
         else:
-            f = open(self.output_file, "w")
+            f = open(self.session.expand_placeholders(self.output_file), "w")
             writer = csv.writer(f)
 
         writer.writerow(["label", "bin", "from", "to", "count"])
@@ -266,7 +267,7 @@ class AreaHistogramWriter(StreamWriter):
         if use_stdout:
             print(json.dumps(data, indent=2))
         else:
-            with open(self.output_file, "w") as f:
+            with open(self.session.expand_placeholders(self.output_file), "w") as f:
                 json.dump(data, f, indent=2)
 
     def output_histograms(self):