+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - switched to underscores in project name
+- added support for placeholders: `area-histogram`, and `label-dist`
 
 
 0.0.1 (2024-05-06)